SERVER-39386 Specifically use the version of the android NDK that we want to use

+++ buildscripts/setup-android-toolchain.sh
@@ -18,19 +18,24 @@ if [ ! -e  $SDK_ROOT ]; then
     (
         cd $SDK_ROOT
         SDK_PACKAGE=sdk-tools-linux-4333796.zip
-        curl -O https://dl.google.com/android/repository/$SDK_PACKAGE
-        unzip $SDK_PACKAGE
+        test -e $SDK_PACKAGE || curl -O https://dl.google.com/android/repository/$SDK_PACKAGE
+        unzip -q $SDK_PACKAGE
         yes | ./tools/bin/sdkmanager --channel=0 \
             "platforms;android-28"  \
-            "ndk-bundle"  \
             "emulator"  \
             "patcher;v4"  \
             "platform-tools"  \
             "build-tools;28.0.0" \
             "system-images;android-21;google_apis;armeabi-v7a"  \
             "system-images;android-24;google_apis;arm64-v8a"  \
-            "system-images;android-21;google_apis;x86_64"
+            "system-images;android-21;google_apis;x86_64" \
+        | grep -v Unzipping
     )
+    NDK=android-ndk-r18b
+    NDK_PACKAGE=$NDK-linux-x86_64.zip
+    test -e $NDK_PACKAGE || curl -O https://dl.google.com/android/repository/$NDK_PACKAGE
+    unzip -q $NDK_PACKAGE
+    mv $NDK $SDK_ROOT/ndk-bundle
 fi
 
 TOOLCHAIN=$PWD/android_toolchain-${ToolchainArch}-${API_VERSION}
